Siirretty CommandLine olio tempDb:ltä embeddedJettyLauncheriin ja tempDb poistettu
